@@ -4,7 +4,7 @@
   <parent>
     <groupId>org.smartregister</groupId>
     <artifactId>opensrp-gateway-plugin</artifactId>
-    <version>2.0.8-alpha</version>
+    <version>2.2.2-alpha</version>
   </parent>
 
   <artifactId>exec</artifactId>
@@ -12,7 +12,7 @@
 
   <properties>
     <root.basedir>${project.parent.basedir}</root.basedir>
-    <spring-boot.version>2.7.5</spring-boot.version>
+    <spring-boot.version>3.3.4</spring-boot.version>
     <!-- We do not want to deploy the `exec` uber jar. -->
     <maven.deploy.skip>true</maven.deploy.skip>
   </properties>
@@ -46,13 +46,13 @@
     <dependency>
       <groupId>com.google.fhir.gateway</groupId>
       <artifactId>server</artifactId>
-      <version>0.3.1</version>
+      <version>${fhir.gateway.version}</version>
     </dependency>
 
     <dependency>
       <groupId>com.google.fhir.gateway</groupId>
       <artifactId>plugins</artifactId>
-      <version>0.3.1</version>
+      <version>${fhir.gateway.version}</version>
     </dependency>
 
     <!-- For Spring -->
@@ -70,7 +70,7 @@
     <dependency>
       <groupId>org.smartregister</groupId>
       <artifactId>plugins</artifactId>
-      <version>2.0.8-alpha</version>
+      <version>2.2.2-alpha</version>
     </dependency>
 
     <dependency>
@@ -98,7 +98,7 @@
             </goals>
             <configuration>
               <mainClass>org.smartregister.fhir.gateway.MainApp</mainClass>
-              <!-- We use ZIP layout such that PropertiesLauncher is used and we
+              <!-- We use ZIP layout such that PropertiesLauncher is used, and we
               can add extra jars through `loader.path` property:
               https://docs.spring.io/spring-boot/docs/current/reference/html/executable-jar.html#appendix.executable-jar.property-launcher
               This is useful for when people want to inject other plugins into
