@@ -4,38 +4,38 @@
   <parent>
     <groupId>org.smartregister</groupId>
     <artifactId>opensrp-gateway-plugin</artifactId>
-    <version>2.0.8-alpha</version>
+    <version>2.2.2-alpha</version>
   </parent>
 
   <artifactId>plugins</artifactId>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
     <project.basedir>${project.basedir}</project.basedir>
-    <hapifhir.version>6.8.3</hapifhir.version>
+    <hapifhir.version>7.4.3</hapifhir.version>
     <fhir.proxy-plugin.version>0.2.0</fhir.proxy-plugin.version>
+    <sentry.version>7.15.0</sentry.version>
+    <jakarta-servlet.version>6.0.0</jakarta-servlet.version>
   </properties>
 
   <dependencies>
-
     <dependency>
       <groupId>com.google.fhir.gateway</groupId>
       <artifactId>server</artifactId>
-      <version>0.3.2</version>
+      <version>${fhir.gateway.version}</version>
     </dependency>
     <dependency>
       <groupId>com.google.fhir.gateway</groupId>
       <artifactId>plugins</artifactId>
-      <version>0.3.1</version>
+      <version>${fhir.gateway.version}</version>
     </dependency>
     <!-- Needed for JEE/Servlet support -->
     <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>4.0.1</version>
-      <scope>provided</scope>
+      <groupId>jakarta.servlet</groupId>
+      <artifactId>jakarta.servlet-api</artifactId>
+      <version>${jakarta-servlet.version}</version>
     </dependency>
     <dependency>
       <groupId>ca.uhn.hapi.fhir</groupId>
@@ -51,12 +51,12 @@
     <dependency>
       <groupId>io.sentry</groupId>
       <artifactId>sentry-logback</artifactId>
-      <version>7.6.0</version>
+      <version>${sentry.version}</version>
     </dependency>
     <dependency>
       <groupId>io.sentry</groupId>
-      <artifactId>sentry-spring-boot-starter</artifactId>
-      <version>7.6.0</version>
+      <artifactId>sentry-spring-boot-starter-jakarta</artifactId>
+      <version>${sentry.version}</version>
     </dependency>
   </dependencies>
 
