@@ -5,11 +5,34 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
-import javax.servlet.http.HttpServletResponse;
-
 import org.junit.Test;
+import org.mockito.Mockito;
+
+import com.google.fhir.gateway.TokenVerifier;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 public class RestUtilTest {
+
+    @Test(expected = RuntimeException.class)
+    public void testCheckAuthenticationThrowsExceptionWhenNoAuthHeader() {
+        HttpServletRequest requestMock = mock(HttpServletRequest.class);
+        TokenVerifier tokenVerifierMock = mock(TokenVerifier.class);
+        Mockito.when(requestMock.getHeader(Constants.AUTHORIZATION)).thenReturn(null);
+        RestUtils.checkAuthentication(requestMock, tokenVerifierMock);
+    }
+
+    @Test
+    public void testCheckAuthenticationCallsTokenVerifierWhenAuthHeaderExists() {
+        HttpServletRequest requestMock = mock(HttpServletRequest.class);
+        TokenVerifier tokenVerifierMock = mock(TokenVerifier.class);
+        String authHeader = "Bearer someToken";
+        Mockito.when(requestMock.getHeader(Constants.AUTHORIZATION)).thenReturn(authHeader);
+        RestUtils.checkAuthentication(requestMock, tokenVerifierMock);
+        verify(tokenVerifierMock).decodeAndVerifyBearerToken(authHeader);
+    }
+
     @Test
     public void testAddCorsHeadersSetsCorsHeaders() {
         HttpServletResponse responseMock = mock(HttpServletResponse.class);
